requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
attrs==20.3.0
Automat==0.6.0
certifi==2020.12.5
chardet==4.0.0
click==6.7
cloud-init==20.1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cssselect==1.1.0
distro-info===0.18ubuntu0.18.04.1
Django==3.2.12
djangorestframework==3.13.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
importlib-metadata==4.8.3
incremental==16.10.1
Jinja2==2.11.3
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==1.1.1
mlnx-tools===-5.2.0-
netifaces==0.10.4
oauthlib==3.1.0
PAM==0.4.2
pexpect==4.2.1
Pillow==8.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycurl==7.43.0.1
PyGObject==3.26.1
PyMySQL==1.0.2
pyOpenSSL==17.5.0
pyquery==1.4.3
pyrsistent==0.17.3
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-debian==0.1.32
pytz==2021.3
pyxdg==0.25
PyYAML==5.4.1
requests==2.25.1
requests-unixsocket==0.1.5
sass==2.3
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
sos==4.3
sqlparse==0.4.2
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing_extensions==4.1.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
zipp==3.4.1
zope.interface==4.3.2
